+++ qcp/gui/.gitignore
