+++ qcp/gui/components/__init__.py
@@ -15,8 +15,8 @@
 Code to handle the behaviour of elements of our UI, split into smaller
 UI components for clarity.
 """
-from qcp.ui.components.abstract_component import AbstractComponent  # noqa: F401, E501
-from qcp.ui.components.graph_component import GraphComponent  # noqa: F401
-from qcp.ui.components.input_component import InputComponent  # noqa: F401
-from qcp.ui.components.button_component import ButtonComponent  # noqa: F401
-from qcp.ui.components.simulator_component import SimulatorComponent  # noqa: F401, E501
+from qcp.gui.components.abstract_component import AbstractComponent  # noqa: F401, E501
+from qcp.gui.components.graph_component import GraphComponent  # noqa: F401
+from qcp.gui.components.input_component import InputComponent  # noqa: F401
+from qcp.gui.components.button_component import ButtonComponent  # noqa: F401
+from qcp.gui.components.simulator_component import SimulatorComponent  # noqa: F401, E501
