+++ qcp/gui/components/abstract_component.py
