+++ qcp/gui/constants.py
