+++ qcp/gui/form.ui
